Task 2: Create the hash_script and its output file
This is very simple implementation. It's written in bash, so in case
you are using zsh, this may not work. The script checks the number of
arguments and in case it is correct, it proceeds to check if the input
is actually a four digit number. If it is, in fact, a four digit number
the script generates the hash using the sha256sum core-util.